UCT plugin: Separate common code in library file
