+++ src/Makefile.am
@@ -25,6 +25,7 @@ libnccl_net_la_LDFLAGS  += $(UCX_LDFLAGS)
 libnccl_net_la_SOURCES  += \
 	ucx_plugin.c \
 	ucx_rma_plugin.c \
+	ucx_uct_lib.c \
 	ucx_uct_plugin.c
 endif
